==> 2_Machine_learning/05_Combine_area_maps.ipynb <==
## Combine all predictions into a single file

! gdal_merge.py -o results/combinedpredictions.tif -co COMPRESS=Deflate -ot UInt16 -n 255 -a_nodata 255 data/district_*_croptype_prediction.tif

## Combine all probabilities into a single file

! gdal_merge.py -o results/combinedprobabilities.tif -co COMPRESS=Deflate -ot UInt16 -n 255 -a_nodata 255 data/district_*_croptype_probabilities.tif
